requirements_groq.txt
requests>=2.25.0
groq>=0.4.0
python-dotenv>=0.19.0